+++ test/pg/get_billing_key.py
@@ -4,9 +4,9 @@ import os
 sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 
 from bootpay_backend import BootpayBackend
+from config import get_active_pg_config
 
-# bootpay = BootpayBackend('59b731f084382614ebf72215', 'WwDv0UjfwFa04wYG0LJZZv1xwraQnlhnHE375n52X0U=')
-bootpay = BootpayBackend('5b8f6a4d396fa665fdc2b5ea', 'rm6EYECr6aroQVG2ntW0A6LpWnkTgP4uQ3H18sDDUYw=')
+bootpay = BootpayBackend(**get_active_pg_config())
 
 token = bootpay.get_access_token() 
 if 'error_code' not in token: